Refactor instance count caching logic and helper class to seperate files
